% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aklhousingprice}
\alias{aklhousingprice}
\title{Auckland housing price}
\format{
A tibble with 8,011 rows and 10 variables:
\itemize{
\item \code{region}: "Auckland"
\item \code{district}: Auckland districts
\item \code{property_address}: Property address
\item \code{lon}: Latitude of the property
\item \code{lat}: Longitude of the property
\item \code{auction_price}: Auction price
\item \code{auction_dates}: Auction date
\item \code{bedrooms}: The number of bedrooms
\item \code{bathrooms}: The number of bathrooms
\item \code{car_parking}: The number of parkings
\item \code{rating_value}: Rating price
\item \code{rating_dates}: Rating dates
}
}
\source{
\href{https://www.interest.co.nz}{interest.co.nz}
}
\usage{
aklhousingprice
}
\description{
Auckland housing price
}
\details{
This data is scraped from the \href{https://www.interest.co.nz}{interest.co.nz}
and contains Auckland auction prices between 2018 and 2021.
}
\keyword{datasets}
