% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiviews.R
\name{hconcat}
\alias{hconcat}
\alias{vconcat}
\title{Concatenate views}
\usage{
hconcat(...)

vconcat(...)
}
\arguments{
\item{...}{A list of \code{vega()} objects.}
}
\description{
\code{hconcat()} for horizontal concatenation, and \code{vconcat()} for vertical
concatenation
}
