% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{enc}
\alias{enc}
\title{Map data variables to visual encodings}
\usage{
enc(x, y, ...)
}
\arguments{
\item{x, y, ...}{A set of name-value pairs to describe the mappings of data
variables to visual encodings in \code{vega()} and individual mark layers \verb{mark_*()}.
Use \code{NULL} to disable a layer encoding to inherit from its parent encodings.}
}
\value{
A list of quosures or constants.
}
\description{
Map data variables to visual encodings
}
\examples{
enc(x = mpg, y = wt)
enc(colour = cyl)
enc(color = cyl)
enc(x = NULL)
}
