% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{encode_if}
\alias{encode_if}
\title{Conditional encoding selection}
\usage{
encode_if(selection, true, false)
}
\arguments{
\item{selection}{A selection or selection compositions.}

\item{true, false}{Values for true/false element of \code{selection}.}
}
\description{
Conditional encoding selection
}
