% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega.R
\name{entitle}
\alias{entitle}
\title{Modify vega title, subtitle, and description}
\usage{
entitle(v, title = NULL, subtitle = NULL, description = NULL)
}
\arguments{
\item{title, subtitle, description}{Strings.}
}
\description{
Modify vega title, subtitle, and description
}
