% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiviews.R
\name{facet_views}
\alias{facet_views}
\title{Facet data views by rows and columns}
\usage{
facet_views(v, row = NULL, column = NULL)
}
\arguments{
\item{v}{A \code{vega()} object.}

\item{row, column}{A set of data variables to define facetted views on the
rows and columns grid.}
}
\description{
Facet data views by rows and columns
}
