% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vegawidget.R, R/vega.R
\name{knit_print.vegaspec}
\alias{knit_print.vegaspec}
\alias{knit_print.virgo}
\title{Knit-print method}
\usage{
knit_print.virgo(spec, ..., renderer = "canvas", options = NULL)
}
\arguments{
\item{spec}{vega spec.}

\item{...}{Options passed to knitr.}

\item{renderer}{One of "svg" or "canvas".}

\item{options}{Options.}
}
\description{
See \code{vegawidget::\link[vegawidget]{knit_print.vegaspec}} for details,
particularly on additional packages that may have to be installed.
}
