% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\alias{config_ggplot}
\title{Vega theme configurations}
\usage{
config(
  v,
  background = "white",
  axis = list(),
  axis_x = list(),
  axis_y = list(),
  header = list(),
  legend = list(),
  title = list(),
  view = list(),
  facet = list()
)

config_ggplot(v)
}
\arguments{
\item{v}{A \code{vega()} object.}

\item{background}{A plot background.}

\item{header, legend, title, view, facet}{A named list.}

\item{axix, axis_x, axis_y}{A named list to define axis.}
}
\description{
Vega theme configurations
}
