% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{vega-dplyr}
\alias{vega-dplyr}
\alias{group_by.virgo_selection}
\alias{mutate.virgo_selection}
\alias{summarise.virgo_selection}
\alias{summarize.virgo_selection}
\title{Dplyr methods for selections}
\usage{
group_by.virgo_selection(.data, ...)

mutate.virgo_selection(.data, ...)

summarise.virgo_selection(.data, ...)

summarize.virgo_selection(.data, ...)
}
\description{
Dplyr methods for selections
}
\keyword{internal}
