% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{input_slider}
\alias{input_slider}
\alias{input_radio}
\alias{input_select}
\alias{input_textbox}
\alias{input_checkbox}
\alias{input_color}
\alias{input_colour}
\alias{input_date}
\alias{input_datetime}
\alias{input_month}
\alias{input_week}
\title{HTML elements that bind to selections}
\usage{
input_slider(name = NULL, min, max, step, init = NULL)

input_radio(name = NULL, choices, init = NULL)

input_select(name = NULL, choices, init = NULL)

input_textbox(init = NULL, ...)

input_checkbox(init = NULL, ...)

input_color(init = NULL, ...)

input_colour(init = NULL, ...)

input_date(init = NULL, ...)

input_datetime(init = NULL, ...)

input_month(init = NULL, ...)

input_week(init = NULL, ...)
}
\arguments{
\item{name}{Name of the HTML input.}

\item{min, max}{Minimum and maximum values.}

\item{step}{Incremental step.}

\item{init}{An initial value.}

\item{choices}{A (named) vector of options.}

\item{...}{Not sure.}
}
\description{
HTML elements that bind to selections
}
