% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_arc}
\alias{mark_arc}
\alias{mark_ribbon}
\alias{mark_boxplot}
\alias{mark_circle}
\alias{mark_errorband}
\alias{mark_image}
\alias{mark_line}
\alias{mark_point}
\alias{mark_rect}
\alias{mark_rule}
\alias{mark_square}
\alias{mark_text}
\alias{mark_tick}
\alias{mark_trail}
\alias{mark_area}
\alias{mark_bar}
\alias{mark_errorbar}
\alias{mark_histogram}
\alias{mark_step}
\alias{mark_density}
\alias{mark_bin2d}
\alias{mark_streamgraph}
\alias{mark_smooth}
\title{Add new marks to \code{vega()} visualisation}
\usage{
mark_arc(v, encoding = NULL, data = NULL, selection = NULL, ..., na.rm = TRUE)

mark_ribbon(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  na.rm = TRUE
)

mark_boxplot(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  na.rm = TRUE
)

mark_circle(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  na.rm = TRUE
)

mark_errorband(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  na.rm = TRUE
)

mark_image(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  na.rm = TRUE
)

mark_line(v, encoding = NULL, data = NULL, selection = NULL, ..., na.rm = TRUE)

mark_point(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  na.rm = TRUE
)

mark_rect(v, encoding = NULL, data = NULL, selection = NULL, ..., na.rm = TRUE)

mark_rule(v, encoding = NULL, data = NULL, selection = NULL, ..., na.rm = TRUE)

mark_square(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  na.rm = TRUE
)

mark_text(v, encoding = NULL, data = NULL, selection = NULL, ..., na.rm = TRUE)

mark_tick(v, encoding = NULL, data = NULL, selection = NULL, ..., na.rm = TRUE)

mark_trail(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  na.rm = TRUE
)

mark_area(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  position = "stack",
  ...,
  na.rm = TRUE
)

mark_bar(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  position = "stack",
  ...,
  na.rm = TRUE
)

mark_errorbar(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  na.rm = TRUE
)

mark_histogram(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  position = "stack",
  ...,
  bin = TRUE,
  na.rm = TRUE
)

mark_step(v, encoding = NULL, data = NULL, selection = NULL, ..., na.rm = TRUE)

mark_density(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  position = "identity",
  ...,
  density = list(),
  na.rm = TRUE
)

mark_bin2d(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  bin = list(x = TRUE, y = TRUE),
  na.rm = TRUE
)

mark_streamgraph(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  na.rm = TRUE
)

mark_smooth(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  method = "lm",
  formula = y ~ x,
  bandwidth = 0.3,
  na.rm = TRUE
)
}
\arguments{
\item{v}{A \code{vega()} object.}

\item{encoding}{An aesthetic mapping via \code{enc()}.}

\item{data}{A data frame for the layer.}

\item{selection}{A selection object.}

\item{...}{Additional mark properties.}

\item{na.rm}{If \code{TRUE}, missing values are removed with a message.
If \code{FALSE}, missing values are included.}

\item{position}{One of "identity", "stack", "fill".}

\item{bin}{A list of \code{bin} parameters.}

\item{density}{Density parameters.}

\item{method}{One of "lm" or "loess".}

\item{formula}{One of:
\itemize{
\item y ~ x
\item y ~ x^2
\item y ~ x^\link{order}
\item y ~ log(x)
\item y ~ exp(x)
}}

\item{bandwidth}{Degree of smoother.}
}
\description{
Add new marks to \code{vega()} visualisation
}
