% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_x}
\alias{scale_x}
\alias{scale_y}
\alias{scale_color}
\alias{scale_colour}
\alias{scale_size}
\alias{scale_shape}
\title{Vega scales}
\usage{
scale_x(
  v,
  name = zap(),
  domain = zap(),
  type = "linear",
  breaks = zap(),
  orient = "bottom",
  ...
)

scale_y(
  v,
  name = zap(),
  domain = zap(),
  type = "linear",
  breaks = zap(),
  orient = "left",
  ...
)

scale_color(v, name = zap(), range = zap(), scheme = zap(), guide = TRUE, ...)

scale_colour(v, name = zap(), range = zap(), scheme = zap(), guide = TRUE, ...)

scale_size(v, name = zap(), range = zap(), type = "linear", guide = TRUE, ...)

scale_shape(v, name = zap(), guide = TRUE, ...)
}
\arguments{
\item{v}{A \code{vega()} object.}

\item{name}{A string for an axis label. \code{zap()} is the default label.
\code{NULL} removes the label.}

\item{domain}{A vector of two elements to define the range.}

\item{type}{One of "linear", "log", "sqrt", "temporal", "band", "category" scale types.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{zap()} for default breaks
\item A vector for custom breaks
}}

\item{orient}{One of "bottom" and "top" for \code{scale_x()}. One of "left" and "right"
for \code{scale_y()}.}

\item{...}{Other parameters passed to vega specs.}

\item{range}{Custom range specification for colour, opacity, and size.}

\item{scheme}{Colour scheme.}

\item{guide}{If \code{FALSE}, remove the legend.}
}
\description{
Vega scales
}
