% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{select_single}
\alias{select_single}
\alias{select_multi}
\alias{select_interval}
\alias{select_legend}
\alias{select_domain}
\alias{select_bind}
\title{Initiate a selection}
\usage{
select_single(
  encodings = NULL,
  fields = NULL,
  init = NULL,
  nearest = FALSE,
  on = "click",
  clear = "dblclick",
  empty = "all",
  resolve = "global"
)

select_multi(
  encodings = NULL,
  fields = NULL,
  init = NULL,
  toggle = TRUE,
  nearest = FALSE,
  on = "click",
  clear = "dblclick",
  empty = "all",
  resolve = "global"
)

select_interval(
  encodings = c("x", "y"),
  init = NULL,
  mark = NULL,
  on = "[mousedown, window:mouseup] > window:mousemove!",
  clear = "dblclick",
  translate = on,
  empty = "all",
  zoom = TRUE,
  resolve = "global"
)

select_legend(fields, on = "click", clear = "dblclick")

select_domain()

select_bind(...)
}
\arguments{
\item{encodings}{A character vector of encoding channels, such as "x" and "y".}

\item{fields}{A character vector of data fields.}

\item{init}{An initial value upon selection.}

\item{nearest}{If \code{FALSE}, data values must be interacted with directly to
be added to the selection.}

\item{on, clear}{An event type that triggers/clears the selection. Options are
"click", "dblclick", "dragenter", "dragleave", "dragover", "keydown", "keypress",
"keyup", "mousedown", "mouseover", "mousemove", "mouseout", "mouseup",
"mousewheel", "touchend", "touchmove", "touchstart", "wheel".}

\item{empty}{An empty selection includes "all" or "none" data values.}

\item{resolve}{One of "global", "union", "intersect" options to resolve
ambiguity for layered and multi-view displays.}

\item{toggle}{A logical to control whether data values should be toggled or
only ever inserted into multi selections.}

\item{mark}{A named vector of mark properties for brushed rectangle.}

\item{translate}{A string or logical to interactively move an interval
selection back-and-forth.}

\item{zoom}{If \code{TRUE}, interactively resize an interval selection.}

\item{...}{A set of name-value pairs with data variables on the LHS and
\verb{input_*()} on the RHS.}
}
\description{
Initiate a selection
}
\section{Composing Multiple Selections}{

A set of operations ...
}

