% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega.R
\name{vega_serialise_data}
\alias{vega_serialise_data}
\alias{vega_serialize_data}
\title{Serialise data}
\usage{
vega_serialise_data(v, path = NULL)

vega_serialize_data(v, path = NULL)
}
\arguments{
\item{v}{A \code{vega()} object.}

\item{path}{Directory to save inlining data to external data files.}
}
\description{
Serialise data
}
