% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega.R
\name{vega_set_size}
\alias{vega_set_size}
\alias{vega_set_width}
\alias{vega_set_height}
\title{Set vega canvas size}
\usage{
vega_set_size(v, width = NULL, height = NULL)

vega_set_width(v, width = NULL)

vega_set_height(v, height = NULL)
}
\arguments{
\item{v}{A \code{vega()} object.}

\item{width, height}{Data plotting width and height.}
}
\description{
Set vega canvas size
}
