% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vega.R
\name{vega}
\alias{vega}
\title{Create a new vega visualisation}
\usage{
vega(data = NULL, encoding = enc(), width = 300, height = 300)
}
\arguments{
\item{data}{A data frame.}

\item{encoding}{A list of aethetic encodings via \code{\link[=enc]{enc()}}.}

\item{width, height}{Data plotting width and height.}
}
\description{
Create a new vega visualisation
}
