% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{vg_sum}
\alias{vg_sum}
\alias{vg_min}
\alias{vg_max}
\alias{vg_mean}
\alias{vg_median}
\alias{vg_count}
\alias{vg_distinct}
\alias{vg_argmin}
\alias{vg_argmax}
\title{Interactive aggregation operations}
\usage{
vg_sum(x)

vg_min(x)

vg_max(x)

vg_mean(x)

vg_median(x)

vg_count(x)

vg_distinct(x)

vg_argmin(x, y)

vg_argmax(x, y)
}
\arguments{
\item{x}{A data variable, used in conjunction with \code{enc()} or dplyr verbs.
\code{vg_count()} can accept an empty input.}

\item{y}{A data variable that maxmises/minimises \code{x} in \code{vg_argmin()} and
\code{vg_argmax()}.}
}
\description{
Interactive aggregation operations
}
