% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{vg_year}
\alias{vg_year}
\alias{vg_quarter}
\alias{vg_month}
\alias{vg_yearmonth}
\alias{vg_date}
\alias{vg_week}
\alias{vg_day}
\alias{vg_dayofyear}
\alias{vg_hours}
\alias{vg_minutes}
\alias{vg_seconds}
\alias{vg_milliseconds}
\title{Interactive time unit operations}
\usage{
vg_year(x, step = 1, utc = FALSE)

vg_quarter(x, step = 1, utc = FALSE)

vg_month(x, step = 1, utc = FALSE)

vg_yearmonth(x, step = 1, utc = FALSE)

vg_date(x, step = 1, utc = FALSE)

vg_week(x, step = 1, utc = FALSE)

vg_day(x, step = 1, utc = FALSE)

vg_dayofyear(x, step = 1, utc = FALSE)

vg_hours(x, step = 1, utc = FALSE)

vg_minutes(x, step = 1, utc = FALSE)

vg_seconds(x, step = 1, utc = FALSE)

vg_milliseconds(x, step = 1, utc = FALSE)
}
\arguments{
\item{x}{A data variable, used in conjunction with \code{enc()} or dplyr verbs.
\code{vg_count()} can accept an empty input.}

\item{step}{An integer to define the number of time steps.}

\item{utc}{If \code{TRUE}, parse data in UTC time, otherwise in local time.}
}
\description{
Interactive time unit operations
}
