% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{vg_window_sum}
\alias{vg_window_sum}
\alias{vg_window_mean}
\alias{vg_window_rank}
\alias{vg_window_count}
\alias{vg_cumsum}
\alias{vg_cummean}
\alias{vg_ranking}
\alias{vg_row_number}
\alias{vg_rank}
\alias{vg_dense_rank}
\alias{vg_percent_rank}
\alias{vg_cume_dist}
\alias{vg_ntile}
\alias{vg_lead}
\alias{vg_lag}
\title{Interactive window operations}
\usage{
vg_window_sum(x, frame = list(NULL, 0), sort = NULL)

vg_window_mean(x, frame = list(NULL, 0), sort = NULL)

vg_window_rank(x, frame = list(NULL, 0), sort = NULL)

vg_window_count(x, frame = list(NULL, 0), sort = NULL)

vg_cumsum(x, sort = NULL)

vg_cummean(x, sort = NULL)

vg_ranking(x, n = 1, sort = NULL, op)

vg_row_number(x, sort = NULL)

vg_rank(x, sort = NULL)

vg_dense_rank(x, sort = NULL)

vg_percent_rank(x, sort = NULL)

vg_cume_dist(x, sort = NULL)

vg_ntile(x, n = 1, sort = NULL)

vg_lead(x, n = 1, sort = NULL)

vg_lag(x, n = 1, sort = NULL)
}
\arguments{
\item{x}{A data variable, used in conjunction with \code{dplyr::mutate()}.}

\item{frame}{A list/vector of two elements to indicate the number of data values
preceding and following the current data object. \code{NULL} gives unbounded elements
proceding or following the current position.}

\item{sort}{A variable for sorting data within a window in ascending order.
\code{-} before the variable gives descending order. \code{NULL} disables sorting.}
}
\description{
Interactive window operations
}
