% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{melbweather}
\alias{melbweather}
\title{Melbourne microclimate measurements}
\format{
A tibble with 43,027 rows and 9 variables:
\itemize{
\item \code{site}: Site identifier, this is the location of the weather sensor,
there are five sites
\item \verb{longitude, latitude}: The spatial coordinates of the measurement sites
\item \code{date_time}: The local date time that a sensor made a recording
\item \code{date}: Date associated with \code{date_time}
\item \code{ambient_temperature}: The value of the ambient air temperature in degrees Celsius.
\item \code{relative_humidity}: The percent value of the relative humidity
\item \code{barometric_pressure}: The barometric pressure in hectopascals (hPa)
\item \code{wind_speed}: The wind speed in kilometers per hour (km/h)
\item \verb{pm2.5,pm10}: The mass density of particulate matter in the air less than 2.5 (10) micrometers in diameter. Measured in micrograms per cubic meter.
}
}
\source{
\href{https://data.melbourne.vic.gov.au/Environment/Microclimate-Sensor-Readings/u4vh-84j8}{Melbourne Open Data Portal}
}
\usage{
melbweather
}
\description{
Melbourne microclimate measurements
}
\details{
This data comes from the \href{https://data.melbourne.vic.gov.au/Environment/Microclimate-Sensor-Readings/u4vh-84j8}{Melbourne Open Data Portal}
and contains measurements from microclimate sensors around the city. Here
we have restricted the data to contain measurements from December 2019 until
the end of Februrary 2020. There are five sites where measurements are taken
every 15 minutes. The data is presented in wide form
}
\keyword{datasets}
