% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark.R
\name{mark_arc}
\alias{mark_arc}
\alias{mark_ribbon}
\alias{mark_boxplot}
\alias{mark_circle}
\alias{mark_errorband}
\alias{mark_image}
\alias{mark_line}
\alias{mark_point}
\alias{mark_rect}
\alias{mark_rule}
\alias{mark_square}
\alias{mark_text}
\alias{mark_tick}
\alias{mark_trail}
\alias{mark_area}
\alias{mark_bar}
\alias{mark_errorbar}
\alias{mark_histogram}
\alias{mark_step}
\alias{mark_density}
\alias{mark_bin2d}
\alias{mark_streamgraph}
\alias{mark_smooth}
\title{Add new marks to \code{vega()} visualisation}
\usage{
mark_arc(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_ribbon(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_boxplot(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_circle(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_errorband(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_image(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_line(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_point(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_rect(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_rule(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_square(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_text(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_tick(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_trail(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_area(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  position = "stack"
)

mark_bar(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  position = "stack"
)

mark_errorbar(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_histogram(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  position = "stack",
  bin = TRUE
)

mark_step(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_density(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  position = "identity",
  density = list()
)

mark_bin2d(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  bin = list(x = TRUE, y = TRUE)
)

mark_streamgraph(v, encoding = NULL, data = NULL, selection = NULL, ...)

mark_smooth(
  v,
  encoding = NULL,
  data = NULL,
  selection = NULL,
  ...,
  method = "lm",
  formula = y ~ x,
  bandwidth = 0.3
)
}
\arguments{
\item{v}{A \code{vega()} object.}

\item{encoding}{An aesthetic mapping via \code{enc()}.}

\item{data}{A data frame for the layer.}

\item{selection}{A selection object.}

\item{...}{Additional mark properties.}
}
\description{
Add new marks to \code{vega()} visualisation
}
